requirements.txt
# Claude API
anthropic>=0.40.0

# Amazon PA-API (primary discovery)

# Fallback scraping
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.3.0

# Configuration
python-dotenv>=1.0.0

# Database / ORM
sqlalchemy>=2.0.0

# Scheduling
APScheduler>=3.10.0

# CLI
click>=8.1.0

# Rate limiting / retry
tenacity>=9.0.0

# Text utilities
python-slugify>=8.0.0
jaconv>=0.3.4

# Web automation (note.com publishing)
playwright>=1.44.0

# Logging
structlog>=24.0.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# Testing
pytest>=8.0.0
pytest-mock>=3.14.0
responses>=0.25.0
